feat: Make ScreenStateConfig composable-driven
Refactor the `ScreenStateConfig` interface to be driven by composable functions for `ProgressDialog` and `InfoDialog`.

This change replaces direct composable function implementations within the interface with abstract function declarations. A new factory function, `screenStateConfig`, is introduced to provide default implementations and allow for easier customization. The `ScreenStateWrapper` is updated to use this new factory function by default.